refactor: 예약 삭제 API adminController로 이동
