+++ src/test/java/roomescape/acceptance/WaitingAcceptanceTest.java
@@ -138,7 +138,7 @@ class WaitingAcceptanceTest extends AcceptanceTest {
                 DynamicTest.dynamicTest("어드민이 guest1의 예약을 삭제한다.", () -> {
                     RestAssured.given().log().all()
                             .cookie("token", adminToken)
-                            .when().delete("/reservations/" + reservationId)
+                            .when().delete("/admin/reservations/" + reservationId)
                             .then().log().all()
                             .assertThat().statusCode(204);
                 }),
@@ -187,7 +187,7 @@ class WaitingAcceptanceTest extends AcceptanceTest {
                 DynamicTest.dynamicTest("어드민이 guest1의 예약을 삭제한다.", () -> {
                     RestAssured.given().log().all()
                             .cookie("token", adminToken)
-                            .when().delete("/reservations/" + reservationId)
+                            .when().delete("/admin/reservations/" + reservationId)
                             .then().log().all()
                             .assertThat().statusCode(204);
                 }),
@@ -245,7 +245,7 @@ class WaitingAcceptanceTest extends AcceptanceTest {
                 DynamicTest.dynamicTest("어드민이 guest1의 예약을 취소한다.", () -> {
                     RestAssured.given().log().all()
                             .cookie("token", adminToken)
-                            .when().delete("/reservations/" + reservationId)
+                            .when().delete("/admin/reservations/" + reservationId)
                             .then().log().all()
                             .assertThat().statusCode(204);
                 }),
